api/requirements.txt
azure-functions
google-genai
